fix: fixed bugs caused by trying to message yourself
bugs:\n1. the chat with yourself wouldn't be shown since there was a condition to return null if cant find the other memeber of private channel\n 2. the create would create two channel members record (which is correct for every private chat other than with yourself) which would cause key duplication in front end and possible future problems in back