fix: fixed sync enigne bugs
bugs: 1.the query would not load latest messages if it was exhausted 2. query wouldn't load latest messages if it had loaded more messages with pagination 3.the query wouldn't be exhausted even if all the messages were in local cache since it hadn't loaded them. this bugs were fixed by replacing the query with two queries one for the latest messages and the other for pagination which isn't the best but it works, result of both qureies is cached and the cache is used when paginaiing to prevenet unnecessary loads. note: the latest result always loads at least 50 (mostly repeated) this is done to ensure no message is lost